backend chatBot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
requests==2.32.3
numpy==2.0.1
scikit-learn==1.5.1
sentence-transformers==3.0.1
PyYAML==6.0.2
openai==0.28.0
pdfminer.six==20231228
